Address Brian's comments
