+++ https/testdata/tls_config_noAuth_noHTTP2.good.yml
@@ -6,5 +6,5 @@ tls_config :
   cipher_suites:
     - TLS_RSA_WITH_AES_128_CBC_SHA
   max_version: TLS12
-http_config:
+http_server_config:
   disable_http2: true
